requirements.txt
flask
mega.py
firebase-admin
gunicorn
